Resolve comments on unit tests
